backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.6
pydantic-settings==2.7.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
httpx==0.28.1
apscheduler==3.10.4
python-multipart==0.0.20
email-validator==2.2.0
